+++ lib/yabeda/dsl/class_methods.rb
@@ -92,6 +92,16 @@ module Yabeda
         Thread.current[:yabeda_temporary_tags] ||= {}
       end
 
+      # Limit all group metrics to specific adapters only
+      #
+      # @param adapter_names [Array<Symbol>] Names of adapters to use
+      def adapter(*adapter_names, group: @group)
+        raise ConfigurationError, "Adapter limitation can't be defined outside of group" unless group
+
+        Yabeda.groups[group] ||= Yabeda::Group.new(group)
+        Yabeda.groups[group].adapter(*adapter_names)
+      end
+
       private
 
       def register_metric(metric)
@@ -99,7 +109,7 @@ module Yabeda
         ::Yabeda.define_singleton_method(name) { metric }
         ::Yabeda.metrics[name] = metric
         register_group_for(metric) if metric.group
-        ::Yabeda.adapters.each_value { |adapter| adapter.register!(metric) } if ::Yabeda.configured?
+        metric.adapters.each_value { |adapter| adapter.register!(metric) } if ::Yabeda.configured?
         metric
       end