+++ lib/yabeda/gauge.rb
@@ -6,7 +6,7 @@ module Yabeda
     def set(tags, value)
       all_tags = ::Yabeda::Tags.build(tags, group)
       values[all_tags] = value
-      ::Yabeda.adapters.each_value do |adapter|
+      adapters.each_value do |adapter|
         adapter.perform_gauge_set!(self, all_tags, value)
       end
       value