+++ lib/yabeda/histogram.rb
@@ -20,7 +20,7 @@ module Yabeda
 
       all_tags = ::Yabeda::Tags.build(tags, group)
       values[all_tags] = value
-      ::Yabeda.adapters.each_value do |adapter|
+      adapters.each_value do |adapter|
         adapter.perform_histogram_measure!(self, all_tags, value)
       end
       value