+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Ability to limit some metrics to specific adapters. [#37](https://github.com/yabeda-rb/yabeda/pull/37) by [@Keallar] and [@Envek]
+
+  ```ruby
+  Yabeda.configure do
+    group :cloud do
+      adapter :newrelic, :datadog
+
+      counter :foo
+    end
+
+    counter :bar, adapter: :prometheus
+  end
+  ```
+
 - Multiple expectations in RSpec matchers:
 
   ```ruby
@@ -155,3 +169,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [@dsalahutdinov]: https://github.com/dsalahutdinov "Dmitry Salahutdinov"
 [@asusikov]: https://github.com/asusikov "Alexander Susikov"
 [@liaden]: https://github.com/liaden "Joel Johnson"
+[@Keallar]: https://github.com/Keallar "Eugene Lysanskiy"